requirements-full.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
torch
transformers
detoxify
protobuf
openai
jinja2
accelerate
websockets
anyio
groq
mangum
httpx